item-1-backend-v1/requirements.txt
langchain==0.0.270
streamlit==1.25.0
streamlit_chat=0.1.1
python-dotenv=1.0.0
pandas=1.5.3
flask=2.2.2
requests=2.31.0
python-telegram-bot=20.4
python-dotenv=1.0.0
asgiref=3.7.2
